Crate dmgrs

Stuff to work with Dmgrs assembly files.

Modules

easy_lexer

Provides an easy to use lexer

header

The ROM header declares a bunch of stuff about the ROM.

lexer

Stuff that “lexes” source text into “lexemes”.

multi_line_comments

A multi-line comment is any amount of stuff between /* and */, nesting allowed.

parser

The “parser” stage turns the stream of lexemes (from the lexer) into ab “abstract syntax tree”.

repeated_newline_filter

Repeated newlines are not needed, it’s better to handle them early.

str_cache_impl

Module with a very simple str caching implementation.

Structs

Section

Enums

Condition

RelativeJumpResolveError

SectionLine

Functions

next_id

Generates a usize id value, used for sticking on the end of made up labels.