Module dmgrs::lexer

Stuff that “lexes” source text into “lexemes”.

Use the DmgToken enum and the Logos trait to lex a string into its tokens.

use dmgrs::lexer::Lexeme;
use logos::Logos; // trait, provides the `lexer` function

const SRC: &str = "ld a, 0";
for (span, lexeme) in Lexeme::lexer(SRC).spanned() {
  println!("{span:?} => {lexeme:?}");
}

Most likely you’ll want to parse the token stream that come out of lexing. You can pass the Lexeme::lexer(SRC).spanned() iterator over to chumsky::Stream::from_iter. You’ll also need to tell chumsky what the “span” value will be when you’ve reached the end-of-input, which will generally be the source string length as both sides of a .. range:

use chumsky::stream::Stream;
use dmgrs::lexer::Lexeme;
use logos::Logos; // trait, provides the `lexer` function

const SRC: &str = "ld a, 0";
let len = SRC.len();
let lex_iter = Lexeme::lexer(SRC).spanned();
let stream = Stream::from_iter(len..len, lex_iter);

Enums

Lexeme

The possible tokens within a Dmgrs program.

Functions

try_bin_str_to_u16

Tries to parse a binary digit string into a u16.

try_dec_str_to_u16

Tries to parse a decimal digit string into a u16.

try_hex_str_to_u16

Tries to parse a hexadecimal digit string into a u16.