Enum dmgrs::lexer::DmgToken

pub enum DmgToken {

    Error,
    HorizontalWhitespace,
    EndOfLineComment,
    StartMultiComment,
    EndMultiComment,
    EndOfLine,
    Punctuation(char),
    Ident(Box<str>),
    BinaryLiteral(u16),
    DecimalLiteral(u16),
    HexLiteral(u16),
    StrLiteral(Box<str>),
    KeywordConst,
    KeywordStatic,
    KeywordFn,
    KeywordIf,
    KeywordElse,
    KeywordLoop,
    KeywordContinue,
    KeywordBreak,
    KeywordReturn,
    InstructionAdc,
    InstructionAdd,
    InstructionAnd,
    InstructionBit,
    InstructionCall,
    InstructionCcf,
    InstructionDaa,
    InstructionDec,
    InstructionDi,
    InstructionEi,
    InstructionHalt,
    InstructionInc,
    InstructionJp,
    InstructionJr,
    InstructionLd,
    InstructionLdh,
    InstructionNop,
    InstructionOr,
    InstructionPop,
    InstructionPush,
    InstructionRes,
    InstructionRet,
    InstructionReti,
    InstructionRl,
    InstructionRla,
    InstructionRlc,
    InstructionRlca,
    InstructionRr,
    InstructionRra,
    InstructionRrc,
    InstructionRrca,
    InstructionRst,
    InstructionSbc,
    InstructionScf,
    InstructionSet,
    InstructionSla,
    InstructionSra,
    InstructionSrl,
    InstructionStop,
    InstructionSub,
    InstructionSwap,
    InstructionXor,

}

The possible tokens within a Dmgrs program.

There’s many no-data variants in this enum. Theoretically these are not needed, they could be captured as Ident values. However, Ident stores the identifier as a Box<str>, and the instructions and keywords are presumed to be quite common, so we give them each their own empty variant to avoid excess allocations.

Variants

Error

An error during lexing.

HorizontalWhitespace

Spaces and tabs, but not line endings.

EndOfLineComment

Starts with // and eats characters until the end of line.

StartMultiComment

Multi-line comments start with /*

EndMultiComment

Multi-line comments end with */

EndOfLine

A line ends with \r\n (windows), \n (unix), or \r (pre-unix mac).

Punctuation(char)

A punctuation character (using the [:punct:] regex class) that does not match any other case.

Ident(Box<str>)

A standard identifier in C-style langs: [_a-zA-Z][_a-zA-Z0-9]*

The lone character _ ends up matching as an Ident rather than a Punctuation.

BinaryLiteral(u16)

The digits of a binary literal (no prefix).

Binary literals are a digit sequence (0-1) prefixed with % or 0b

Use of _ is allowed in the literal, but they don’t change the value.

DecimalLiteral(u16)

The digits of a decimal literal.

Decimal literals are a digit sequence (0-9) with no prefix.

Use of _ is allowed in the literal, but they don’t change the value.

HexLiteral(u16)

The digits of a hex literal (no prefix).

Hex literals are a digit sequence (0-9a-fA-F) prefixed with $ or 0x

Use of _ is allowed in the literal, but they don’t change the value.

StrLiteral(Box<str>)

Holds all the stuff between ".

This allows \" and \\ within the string literal that’s collected, but doesn’t actually handle escape sequence processing.

Thanks to Quirl, who made this regex: “works by specifying all of the escape sequences you allow (here just \", and \\ for a \ itself) up front and then requiring the rest of the literal to not be a quote or escape (that would start with \)”

KeywordConst

const

KeywordStatic

static

KeywordFn

fn

KeywordIf

if

KeywordElse

else

KeywordLoop

loop

KeywordContinue

continue

KeywordBreak

break

KeywordReturn

return

InstructionAdc

adc or ADC

InstructionAdd

add or ADD

InstructionAnd

and or AND

InstructionBit

bit or BIT

InstructionCall

call or CALL

InstructionCcf

ccf or CCF

InstructionDaa

daa or DAA

InstructionDec

dec or DEC

InstructionDi

di or DI

InstructionEi

ei or EI

InstructionHalt

halt or HALT

InstructionInc

inc or INC

InstructionJp

jp or JP

InstructionJr

jr or JR

InstructionLd

ld or LD

InstructionLdh

ldh or LDH

InstructionNop

nop or NOP

InstructionOr

or or OR

InstructionPop

pop or POP

InstructionPush

push or PUSH

InstructionRes

res or RES

InstructionRet

ret or RET

InstructionReti

reti or RETI

InstructionRl

rl or RL

InstructionRla

rla or RLA

InstructionRlc

rlc or RLC

InstructionRlca

rlca or RLCA

InstructionRr

rr or RR

InstructionRra

rra or RRA

InstructionRrc

rrc or RRC

InstructionRrca

rrca or RRCA

InstructionRst

rst or RST

InstructionSbc

sbc or SBC

InstructionScf

scf or SCF

InstructionSet

set or SET

InstructionSla

sla or SLA

InstructionSra

sra or SRA

InstructionSrl

srl or SRL

InstructionStop

stop or STOP

InstructionSub

sub or SUB

InstructionSwap

swap or SWAP

InstructionXor

xor or XOR

Implementations

impl DmgToken

pub const fn get_number(&self) -> Option<u16>

Gets the number out of the lexeme, if any.

use DmgToken::*;
assert_eq!(DecimalLiteral(123).get_number(), Some(123));
assert_eq!(InstructionXor.get_number(), None);