Module dmgrs::parser

The “parser” stage turns the stream of lexemes (from the lexer) into ab “abstract syntax tree”.

There is some validation when going from lexeme stream to AST, but not all of it.

Structs

Ast

The Abstract Syntax Tree of a Dmgrs program.

Block

A block of statements

Const

A constant declaration: const FOO = EXPR;

Fn

A function in the program.

Static

A static declaration: static FOO: [u8] = [EXPR, EXPR, EXPR, ...];

Enums

Attribute

Attributes for the program or for a specific fn or static.

BranchTgt

Where to branch the loop to

Expr

An “expression” means an integer constant expression.

InstrArg

An argument to an instruction.

Item

An item is anything that can appear at the top level of a source file.

Stmt

A single statement within a block.