Function dmgrs::str_cache_impl::cache_str

pub fn cache_str(s: &str) -> &'static str

Turns any &str into a &'static str using a minimal number of new allocations.

This works by utilizing a global set of interned str values. When a request comes in for a str to be cached it is looked up in the set. If the str is not already in the set then it’s automatically added to the set. Then a reference to the str in the set is returned to you.

The primary limitation of this system is that it’s only able to grow. It cannot shrink the intern set, since there’s no way to know which str values might still be in use. However, it’s generally sufficient for short lived programs.