Enum dmgrs::lexer::DmgToken

pub enum DmgToken {
    Error,
    HorizontalWhitespace,
    EndOfLineComment,
    StartMultiComment,
    EndMultiComment,
    EndOfLine,
    Punctuation(char),
    Ident(&'static str),
    StrLiteral(&'static str),
    BinaryLiteral(u16),
    DecimalLiteral(u16),
    HexLiteral(u16),
}

The possible tokens within a Dmgrs program.

At this early stage of the processing all keywords, instructions, and register names just count as an “ident”. A simple string interning mechanism is used, so it’s not a super allocation fest to do this.

Variants

Error

An error during lexing.

HorizontalWhitespace

Spaces and tabs, but not line endings.

EndOfLineComment

Starts with // and eats characters until the end of line.

StartMultiComment

Multi-line comments start with /*

EndMultiComment

Multi-line comments end with */

EndOfLine

A line ends with \r\n (windows), \n (unix), or \r (pre-unix mac).

Punctuation(char)

A punctuation character (using the [:punct:] regex class) that does not match any other case.

Ident(&'static str)

A standard identifier in C-style langs: [_a-zA-Z][_a-zA-Z0-9]*

The lone character _ ends up matching as an Ident rather than a Punctuation.

StrLiteral(&'static str)

Holds all the stuff between ".

This allows \" and \\ within the string literal that’s collected, but doesn’t actually handle escape sequence processing.

Thanks to Quirl, who made this regex: “works by specifying all of the escape sequences you allow (here just \", and \\ for a \ itself) up front and then requiring the rest of the literal to not be a quote or escape (that would start with \)”

BinaryLiteral(u16)

The digits of a binary literal (no prefix).

Binary literals are a digit sequence (0-1) prefixed with % or 0b

Use of _ is allowed in the literal, but they don’t change the value.

DecimalLiteral(u16)

The digits of a decimal literal.

Decimal literals are a digit sequence (0-9) with no prefix.

Use of _ is allowed in the literal, but they don’t change the value.

HexLiteral(u16)

The digits of a hex literal (no prefix).

Hex literals are a digit sequence (0-9a-fA-F) prefixed with $ or 0x

Use of _ is allowed in the literal, but they don’t change the value.

Implementations

impl DmgToken

pub const fn get_number(&self) -> Option<u16>

Gets the number out of the lexeme, if any.

use DmgToken::*;
assert_eq!(DecimalLiteral(123).get_number(), Some(123));
assert_eq!(Punctuation('.').get_number(), None);